requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
av==16.1.0
black==24.4.2
blis==0.7.11
braceexpand==0.1.7
cairocffi==1.7.1
CairoSVG==2.8.2
catalogue==2.0.10
certifi==2024.6.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
conda-pack @ file:///croot/conda-pack_1752674263201/work
confection==0.1.5
contourpy==1.2.1
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.8
datasets==4.4.2
decord==0.6.0
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.1
fairscale==0.4.13
faiss-gpu==1.7.2
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
flash_attn==2.8.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.8.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gradio==4.37.1
gradio_client==1.0.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hf-xet==1.1.9
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.34.4
idna==3.7
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
numpy==1.26.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.35.7
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.0
preshed==3.0.9
propcache==0.4.1
protobuf==4.25.3
psutil==7.0.0
pyarrow==22.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.23
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qwen-vl-utils==0.0.14
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.5.0
safetensors==0.4.3
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.35.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-text==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
thinc==8.2.5
timm==1.0.7
tinycss2==1.5.1
tokenizers==0.22.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.4
transformers==4.57.3
triton==3.4.0
typer==0.12.3
typing_extensions==4.15.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uv==0.9.25
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.18.7
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.6.0
yarl==1.22.0
